// packages/cloudformation/generated/resources/aws-lambda-function.ts
export type FunctionProperties = {
  Architectures?: CloudFormationValue<Array<CloudFormationValue<string>>>;
  CapacityProviderConfig?: CloudFormationValue<CapacityProviderConfig>;
  Code: CloudFormationValue<Code>;
  CodeSigningConfigArn?: CloudFormationValue<string>;
  DeadLetterConfig?: CloudFormationValue<DeadLetterConfig>;
  Description?: CloudFormationValue<string>;
  DurableConfig?: CloudFormationValue<DurableConfig>;
  Environment?: CloudFormationValue<Environment>;
  EphemeralStorage?: CloudFormationValue<EphemeralStorage>;
  FileSystemConfigs?: CloudFormationValue<Array<CloudFormationValue<FileSystemConfig>>>;
  FunctionName?: CloudFormationValue<string>;
  FunctionScalingConfig?: CloudFormationValue<FunctionScalingConfig>;
  Handler?: CloudFormationValue<string>;
  ImageConfig?: CloudFormationValue<ImageConfig>;
  KmsKeyArn?: CloudFormationValue<string>;
  Layers?: CloudFormationValue<Array<CloudFormationValue<string>>>;
  LoggingConfig?: CloudFormationValue<LoggingConfig>;
  MemorySize?: CloudFormationValue<number>;
  PackageType?: CloudFormationValue<"Image" | "Zip">;
  PublishToLatestPublished?: CloudFormationValue<boolean>;
  RecursiveLoop?: CloudFormationValue<"Allow" | "Terminate">;
  ReservedConcurrentExecutions?: CloudFormationValue<number>;
  Role: CloudFormationValue<string>;
  Runtime?: CloudFormationValue<string>;
  RuntimeManagementConfig?: CloudFormationValue<RuntimeManagementConfig>;
  SnapStart?: CloudFormationValue<SnapStart>;
  Tags?: CloudFormationValue<Array<CloudFormationValue<CloudFormationTag>>>;
  TenancyConfig?: CloudFormationValue<TenancyConfig>;
  Timeout?: CloudFormationValue<number>;
  TracingConfig?: CloudFormationValue<TracingConfig>;
  VpcConfig?: CloudFormationValue<VpcConfig>;
};

export type CapacityProviderConfig = {
  LambdaManagedInstancesCapacityProviderConfig: CloudFormationValue<LambdaManagedInstancesCapacityProviderConfig>;
};

export type Code = {
  ImageUri?: CloudFormationValue<string>;
  S3Bucket?: CloudFormationValue<string>;
  S3Key?: CloudFormationValue<string>;
  S3ObjectStorageMode?: CloudFormationValue<"COPY" | "REFERENCE">;
  S3ObjectVersion?: CloudFormationValue<string>;
  SourceKMSKeyArn?: CloudFormationValue<string>;
  ZipFile?: CloudFormationValue<string>;
};

export type DeadLetterConfig = {
  TargetArn?: CloudFormationValue<string>;
};

export type DurableConfig = {
  ExecutionTimeout: CloudFormationValue<number>;
  KMSKeyArn?: CloudFormationValue<string>;
  RetentionPeriodInDays?: CloudFormationValue<number>;
};

export type Environment = {
  Variables?: CloudFormationValue<Record<string, CloudFormationValue<string>>>;
};

export type EphemeralStorage = {
  Size: CloudFormationValue<number>;
};

export type FileSystemConfig = {
  Arn: CloudFormationValue<string>;
  LocalMountPath: CloudFormationValue<string>;
};

export type FunctionScalingConfig = {
  MaxExecutionEnvironments?: CloudFormationValue<number>;
  MinExecutionEnvironments?: CloudFormationValue<number>;
};

export type ImageConfig = {
  Command?: CloudFormationValue<Array<CloudFormationValue<string>>>;
  EntryPoint?: CloudFormationValue<Array<CloudFormationValue<string>>>;
  WorkingDirectory?: CloudFormationValue<string>;
};

export type LambdaManagedInstancesCapacityProviderConfig = {
  CapacityProviderArn: CloudFormationValue<string>;
  ExecutionEnvironmentMemoryGiBPerVCpu?: CloudFormationValue<number>;
  PerExecutionEnvironmentMaxConcurrency?: CloudFormationValue<number>;
};

export type LoggingConfig = {
  ApplicationLogLevel?: CloudFormationValue<"TRACE" | "DEBUG" | "INFO" | "WARN" | "ERROR" | "FATAL">;
  LogFormat?: CloudFormationValue<"Text" | "JSON">;
  LogGroup?: CloudFormationValue<string>;
  SystemLogLevel?: CloudFormationValue<"DEBUG" | "INFO" | "WARN">;
};

export type RuntimeManagementConfig = {
  RuntimeVersionArn?: CloudFormationValue<string>;
  UpdateRuntimeOn: CloudFormationValue<"Auto" | "FunctionUpdate" | "Manual">;
};

export type SnapStart = {
  ApplyOn: CloudFormationValue<"PublishedVersions" | "None">;
};

export type TenancyConfig = {
  TenantIsolationMode: CloudFormationValue<"PER_TENANT">;
};

export type TracingConfig = {
  Mode?: CloudFormationValue<"Active" | "PassThrough">;
};

export type VpcConfig = {
  Ipv6AllowedForDualStack?: CloudFormationValue<boolean>;
  SecurityGroupIds?: CloudFormationValue<Array<CloudFormationValue<string>>>;
  SubnetIds?: CloudFormationValue<Array<CloudFormationValue<string>>>;
};